requirements.txt
evdev==1.4.0
RPi.GPIO==0.6.5
